relocate App to app directory, update imports
