+++ src/components/app/types.ts
@@ -1,0 +1,7 @@
+export type Section = {
+  id: "home" | "about" | "skills" | "contact";
+};
+
+export type SectionResponse = {
+  sections: Section[];
+};
