+++ src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
 import "src/global/global.css";
-import App from "src/App";
+import App from "src/components/app/App";
 
 // import i18n (needs to be bundled ;))
 import "src/i18n";
